refactor: split reducer into smaller files
