+++ packages/core/reducer/actions/__tests__/set.spec.ts
@@ -1,0 +1,24 @@
+import { SetDataAction } from "../..";
+import { defaultData, defaultState, testSetup } from "../__helpers__";
+
+describe("Reducer", () => {
+  const { reducer } = testSetup();
+
+  describe("set action", () => {
+    it("should set new data", () => {
+      const newData = {
+        ...defaultData,
+        root: { props: { title: "Hello, world" } },
+        content: [{ type: "Comp", props: { id: "1" } }],
+      };
+
+      const action: SetDataAction = {
+        type: "setData",
+        data: newData,
+      };
+
+      const newState = reducer(defaultState, action);
+      expect(newState.data).toEqual(newData);
+    });
+  });
+});
