+++ packages/core/reducer/index.ts
@@ -1,13 +1,25 @@
 import { Reducer } from "react";
-import { AppState, Config, Data } from "../types";
+import { AppState, Data } from "../types";
 import { PuckAction } from "./actions";
 import type { OnAction } from "../types";
 import { AppStore } from "../store";
 import { PrivateAppState } from "../types/Internal";
-import { reduce } from "./reduce";
+import { setAction } from "./actions/set";
+import { insertAction } from "./actions/insert";
+import { replaceAction } from "./actions/replace";
+import { replaceRootAction } from "./actions/replace-root";
+import { duplicateAction } from "./actions/duplicate";
+import { reorderAction } from "./actions/reorder";
+import { moveAction } from "./actions/move";
+import { removeAction } from "./actions/remove";
+import {
+  registerZoneAction,
+  unregisterZoneAction,
+} from "./actions/register-zone";
+import { setDataAction } from "./actions/set-data";
+import { setUiAction } from "./actions/set-ui";
 
 export * from "./actions";
-export * from "./reduce";
 
 export type ActionType = "insert" | "reorder";
 
@@ -60,9 +72,55 @@ export function createReducer<UserData extends Data>({
 }): StateReducer<UserData> {
   return storeInterceptor(
     (state, action) => {
-      const result = reduce(state, action, appStore);
+      if (action.type === "set") {
+        return setAction(state, action, appStore) as PrivateAppState<UserData>;
+      }
 
-      return result;
+      if (action.type === "insert") {
+        return insertAction(state, action, appStore);
+      }
+
+      if (action.type === "replace") {
+        return replaceAction(state, action, appStore);
+      }
+
+      if (action.type === "replaceRoot") {
+        return replaceRootAction(state, action, appStore);
+      }
+
+      if (action.type === "duplicate") {
+        return duplicateAction(state, action, appStore);
+      }
+
+      if (action.type === "reorder") {
+        return reorderAction(state, action, appStore);
+      }
+
+      if (action.type === "move") {
+        return moveAction(state, action, appStore);
+      }
+
+      if (action.type === "remove") {
+        return removeAction(state, action, appStore);
+      }
+
+      if (action.type === "registerZone") {
+        return registerZoneAction(state, action);
+      }
+
+      if (action.type === "unregisterZone") {
+        return unregisterZoneAction(state, action);
+      }
+
+      if (action.type === "setData") {
+        return setDataAction(state, action, appStore);
+      }
+
+      if (action.type === "setUi") {
+        return setUiAction(state, action);
+      }
+
+      return state;
     },
     record,
     onAction
